create utils folder; move all auth stuff into useAuth.js
