+++ utils/useAuth.js
@@ -1,0 +1,66 @@
+import { useEffect, useState } from "react";
+import netlifyIdentity from "netlify-identity-widget";
+
+export const netlifyAuth = {
+  isAuthenticated: false,
+  user: null,
+  initialize(callback) {
+    window.netlifyIdentity = netlifyIdentity;
+    netlifyIdentity.on("init", (user) => {
+      callback(user);
+    });
+    netlifyIdentity.init();
+  },
+  authenticate(callback) {
+    this.isAuthenticated = true;
+    netlifyIdentity.open();
+    netlifyIdentity.on("login", (user) => {
+      this.user = user;
+      callback(user);
+      netlifyIdentity.close();
+    });
+  },
+  signout(callback) {
+    this.isAuthenticated = false;
+    netlifyIdentity.logout();
+    netlifyIdentity.on("logout", () => {
+      this.user = null;
+      callback();
+    });
+  },
+};
+
+export function useAuth() {
+  let [loggedIn, setLoggedIn] = useState(netlifyAuth.isAuthenticated);
+  let [user, setUser] = useState(null);
+
+  useEffect(() => {
+    let isCurrent = true;
+    netlifyAuth.initialize((user) => {
+      if (isCurrent) {
+        setLoggedIn(!!user);
+        setUser(user);
+      }
+    });
+
+    return () => {
+      isCurrent = false;
+    };
+  }, []);
+
+  let login = () => {
+    netlifyAuth.authenticate((user) => {
+      setLoggedIn(!!user);
+      setUser(user);
+    });
+  };
+
+  let logout = () => {
+    netlifyAuth.signout(() => {
+      setLoggedIn(false);
+      setUser(null);
+    });
+  };
+
+  return { login, logout, user, loggedIn };
+}
